# swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/user":{"post":{"operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OkResponseUser"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseBadRequest"}}}},"404":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseBadRequest"}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseBadRequest"}}}}},"security":[{"bearer":[]}],"tags":["User"]},"get":{"operationId":"UserController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OkResponseUser"}}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseBadRequest"}}}},"404":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseBadRequest"}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseBadRequest"}}}}},"security":[{"bearer":[]}],"tags":["User"]}},"/user/{id}":{"get":{"operationId":"UserController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OkResponseUser"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseBadRequest"}}}},"404":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseBadRequest"}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseBadRequest"}}}}},"security":[{"bearer":[]}],"tags":["User"]},"patch":{"operationId":"UserController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OkResponseUser"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseBadRequest"}}}},"404":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseBadRequest"}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseBadRequest"}}}}},"security":[{"bearer":[]}],"tags":["User"]},"delete":{"operationId":"UserController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OkResponseUser"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseBadRequest"}}}},"404":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseBadRequest"}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseBadRequest"}}}}},"security":[{"bearer":[]}],"tags":["User"]}},"/task":{"post":{"operationId":"TaskController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTaskDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskOkResponseOpenApi"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseBadRequest"}}}},"404":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseBadRequest"}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseBadRequest"}}}}},"security":[{"bearer":[]}],"tags":["Task"]},"get":{"operationId":"TaskController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TaskOkResponseOpenApi"}}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseBadRequest"}}}},"404":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseBadRequest"}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseBadRequest"}}}}},"security":[{"bearer":[]}],"tags":["Task"]}},"/task/{id}":{"get":{"operationId":"TaskController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskOkResponseOpenApi"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseBadRequest"}}}},"404":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseBadRequest"}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseBadRequest"}}}}},"security":[{"bearer":[]}],"tags":["Task"]},"patch":{"operationId":"TaskController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTaskDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskOkResponseOpenApi"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseBadRequest"}}}},"404":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseBadRequest"}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseBadRequest"}}}}},"security":[{"bearer":[]}],"tags":["Task"]},"delete":{"operationId":"TaskController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskOkResponseOpenApi"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseBadRequest"}}}},"404":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseBadRequest"}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseBadRequest"}}}}},"security":[{"bearer":[]}],"tags":["Task"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAuthDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ILoginResData"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseBadRequest"}}}},"404":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseBadRequest"}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseBadRequest"}}}}},"tags":["Auth"]}}},"info":{"title":"TODO-list project","description":"The cats API description","version":"1.0","contact":{}},"tags":[{"name":"api","description":""}],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateUserDto":{"type":"object","properties":{"user":{"type":"string"},"password":{"type":"string"},"refpassword":{"type":"string"},"role":{"type":"string","enum":["admin","worker"]}},"required":["user","password","refpassword","role"]},"User":{"type":"object","properties":{"id":{"type":"number"},"user":{"type":"string"},"password":{"type":"string"},"created_at":{"type":"string"},"updated_at":{"type":"string"},"role":{"type":"string"},"task":{"type":"array","items":{"type":"string"}}},"required":["id","user","password","created_at","updated_at","role","task"]},"BaseMetaDataApi":{"type":"object","properties":{"status":{"type":"number"},"message":{"type":"string"}},"required":["status","message"]},"OkResponseUser":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/User"},"meta":{"$ref":"#/components/schemas/BaseMetaDataApi"}},"required":["data","meta"]},"BaseBadRequest":{"type":"object","properties":{"meta":{"$ref":"#/components/schemas/BaseMetaDataApi"},"data":{"type":"null"}},"required":["meta","data"]},"CreateTaskDto":{"type":"object","properties":{"title":{"type":"string"},"status":{"type":"string","enum":["started","completed","completed"]},"douser":{"type":"number"},"docreator":{"type":"number"}},"required":["title","status","douser","docreator"]},"Task":{"type":"object","properties":{"id":{"type":"number"},"title":{"type":"string"},"status":{"type":"string"},"createat":{"type":"string"},"updateat":{"type":"string"},"creator":{"$ref":"#/components/schemas/User"},"user":{"$ref":"#/components/schemas/User"}},"required":["id","title","status","createat","updateat","creator","user"]},"TaskOkResponseOpenApi":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/Task"},"meta":{"$ref":"#/components/schemas/BaseMetaDataApi"}},"required":["data","meta"]},"UpdateTaskDto":{"type":"object","properties":{"title":{"type":"string"},"status":{"type":"string","enum":["started","completed","completed"]},"douser":{"type":"number"},"docreator":{"type":"number"}}},"CreateAuthDto":{"type":"object","properties":{"user":{"type":"string"},"password":{"type":"string"}},"required":["user","password"]},"IloginType":{"type":"object","properties":{"acc_token":{"type":"string"},"ref_token":{"type":"string"},"user":{"$ref":"#/components/schemas/User"}},"required":["acc_token","ref_token","user"]},"IloginMeta":{"type":"object","properties":{"status":{"type":"number"},"message":{"type":"string"}},"required":["status","message"]},"ILoginResData":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/IloginType"},"meta":{"$ref":"#/components/schemas/IloginMeta"}},"required":["data","meta"]}}}}